+++ Assets/CommonCore/Core/AddonSupport/AddonBase.cs
@@ -12,29 +12,16 @@ namespace CommonCore
         {
             InitialLoadContext = context;
             LocalResourcePath = context.MountPath;
+            context.AddonManager.RegisterDefaultVideoPaths(context);
             await context.AddonManager.LoadAssembliesAsync(context);
             await context.AddonManager.LoadResourcesFromPathAsync(context);
             context.AddonManager.RegisterLoadedScenes(context);
             RunOnLoadedCallback(context);
         }
 
-        //TODO lifecycle events etc?
-
-        [Obsolete]
-        protected async Task LoadAssemblies(AddonLoadContext context)
-        {
-            await context.AddonManager.LoadAssembliesAsync(context);
-        }
-
-        [Obsolete]
-        protected async Task LoadResources(AddonLoadContext context)
-        {
-            await context.AddonManager.LoadResourcesFromPathAsync(context);
-        }
-
         protected void RunOnLoadedCallback(AddonLoadContext context)
         {
-            context.OnLoadedCallback(new AddonLoadData(context.LoadedAssemblies, context.LoadedResources));
+            context.OnLoadedCallback(new AddonLoadData(context.LoadedAssemblies, context.LoadedResources, context.VideoPaths));
         }
     }
 }