+++ CHANGELOG.md
@@ -161,6 +161,8 @@
 * Implemented Generic Translate Script
 * Added flag to pre-activate entity placeholders in world scenes
 * Added UI sound type (PlayUISound now uses this sound type)
+* Removed obsolete methods from AddonBase
+* Added VideoModule to provide addon-aware video management
 * Weapon ViewModel can be forced to always or never wait for lock time to do effect [Experimental]
 * Added subtitle convenience methods from Shattered 2 [Experimental]
 * Added fade canvasgroup convenience methods [Experimental]